+++ lib/bristow/agencies/supervisor.rb
@@ -5,13 +5,15 @@ module Bristow
 
       sgetter :custom_instructions, default: nil
 
-      def initialize(agents: self.class.agents.dup, custom_instructions: self.class.custom_instructions)
+      def initialize(agents: self.class.agents.dup, custom_instructions: self.class.custom_instructions, termination: Bristow::Terminations::MaxMessages.new(100))
         @custom_instructions = custom_instructions
         @agents = agents
+        @termination = termination
         @supervisor = Agents::Supervisor.new(
           child_agents: agents,
           agency: self,
-          custom_instructions: custom_instructions
+          custom_instructions: custom_instructions,
+          termination: termination
         )
       end