+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Add termination conditions so you can ensure an agent stop after a certain amount of time.
 - Drop VCR in favor of WebMock for testing
 
 ## [0.4.0] - 2025-02-15